Add support for Two Factor login
Adds support for TOTP based MFA login. Support is preliminary, as it doesn't adhere to best practices (reset password will also reset MFA), but it adds the options to fix this later, and obviously it's better than not having this feature.

Main implementation details:

1. MFA Details are stored in a separate table linked to the User object. This includes the relevant details to store in a generic enough way to later support alternative providers (email, bypass codes, device, etc).
2. MFA step is handled by sending a semi-valid JWT, which doesn't contain a `userId`, but includes information that the first factor was passed.
3. Only after the second factor is validated will the JWT be upgraded to one which contains the `userId` which the code checks for a fully valid login

There is support for setting up multiple providers (although all of them are TOTP/Authenticator App type only), so people could swap their phone by adding a new one first before removing the old one.

The main issue with the current setup is that it is not considered good practive to allow a user to reset their MFA, but lack of other alternatives (like backup codes or an admin interface where admins can manually reset this) this should be okay.

Roadmap for future improvements:
1. Add support for bypass codes, and automatic bypass code setup once you add your first TOTP MFA
2. Add support for managing User's MFA details through the admin interface
3. Add config options so the 'MFA reset on Password Reset' can be enabled/disabled based on what the instance operator wants

This should cover the main security bits and make sure this implementation is up to current standards

Afterwards:

4. Add support for alternative providers:
4/A. Email might be a good candidate (although it's not really considered secure but some instance providers might enable it for convenience)
4/B. Passkey support would definitely be nice as well
4/C. And finally we might want to use the Wafrn-RN app as well
5. A bit nicer UI, but as I'm mostly a backend dev this will likely remain a dream :D