Refactor user type checking methods
We have some business logic around determining if a user is local, remote, bluesky or fedi. These checks usually determine these based on the presence of '@' in the string with a couple of other checks. Since these checks are scattered around the codebase it is becoming hard to refactor these features if there would be a need - for example if we have to introduce a new type of user. This refactors the backend code to move all of these logix into the `User` class in one place.