Force the usage of package-lock.json in builds
Due to us using a multi-workspace build the package-lock.json was not part of the docker context, meaning it wasn't used when it built the packages internally. This could lead to supply side attacks, where we accidentally download a later - offending - package version during docker builds

Had to update some packages to make sure backend and frontend are in sync.

Note: requires the docker compose files to be overridden. I'll see I can make the update script more clever to get that sorted