Update updater to notify of some breaking changes
This updates the updater so it will notify if anything odd is hapening with either the docker compose or environment exmple files. These are usually the main breaking changes people might need to be aware so hopefully covers the improtant bits.

It will also update the docker compose file autoamtically if it can.

Additional changes are some updates in the docs both for this feature and the one before