Move build arguments to runtime environmental variables
This will allow docker containers to be pre-built and pushed to a docker repository. All of the env var magic is now done during every startup for both the backend and frontend, so pre-build containers can be started up with the env vars set up.

Since angular requires some work the `ngsw.json` config files is handled a bit differently to others.

Some bits will still require a build though:
* Frontend overrides. This is likely not really possible for overriding the frontend itself, but volume mounts should make it okayish for assets, although cache bursting will need some entrypoint love
* PGWatch monitoring. This is likely doable easily but I'm lazy right now to do that ATM